+++ package/thunderstore/Compass/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.2
+* fixed last death pin wasn't showing
+* default maximum distance increased to 550
+* new config option to hold key to see shouts at any distance
+* new config option to hold key to see other players at any distance
+* new config option to hold key to see pings at any distance
+* new config option to hold key to see pin text
+* shouts, players and pings always show its text if any distance key is hold
+
 # 1.0.1
  * Fixed possible error after world relaunch